requirements.txt
configparser==4.0.2
contextlib2==0.6.0.post1
enum34==1.1.6
Flask==1.1.1
functools32==3.2.3.post2
google-api-core==1.31.5
google-api-python-client==1.12.10
google-auth==1.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
h5py==2.10.0
httplib2==0.20.4
importlib-metadata==1.5.0
isort==4.3.21
logging==0.4.9.6
lxml==4.5.2
numpy==1.16.6
oauth2client==4.1.3
oauthlib==3.1.0
passlib==1.7.2
pathlib2==2.3.5
pip-review==1.3.0
protobuf==3.11.3
pyagents==0.5
pycparser==2.19
pylint==1.9.5
PyNaCl==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3
requests==2.18.4
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
uritemplate==3.0.1
urllib3==1.22
virtualenv==16.7.9

